optimize font building
