+++ .gitignore
@@ -12,6 +12,7 @@ stash
 coverage
 stats
 *.log
+outlined
 packages/**/src/icons/*.js
 packages/**/src/icons/*.ts
 packages/**/LICENSE
