+++ package.json
@@ -13,7 +13,7 @@
     "lucide-svelte": "pnpm --filter lucide-svelte",
     "lucide-static": "pnpm --filter lucide-static",
     "build:icons": "node ./scripts/buildIcons.mjs --templateSrc ./packages/lucide/scripts/exportTemplate.mjs",
-    "build:outline-icons": "node ./scripts/outlineSvg.mjs",
+    "build:outline-icons": "pnpm --filter outline-svg start",
     "generate:supersprite": "node ./scripts/generateSuperSVG.mjs",
     "optimize": "node ./scripts/optimizeSvgs.mjs",
     "addtags": "node ./scripts/addMissingKeysToTags.mjs",
@@ -30,6 +30,7 @@
     "@rollup/plugin-commonjs": "^17.1.0",
     "@rollup/plugin-node-resolve": "^11.2.1",
     "@rollup/plugin-replace": "^2.4.2",
+    "@rollup/plugin-typescript": "^8.3.4",
     "babel-jest": "^26.6.3",
     "babel-plugin-add-import-extension": "^1.6.0",
     "core-js": "^3.24.0",
@@ -50,8 +51,6 @@
     "rollup-plugin-svelte": "^7.1.0",
     "rollup-plugin-terser": "^5.3.1",
     "rollup-plugin-visualizer": "^4.2.2",
-    "@rollup/plugin-typescript": "^8.3.4",
-    "svg-outline-stroke": "^1.3.1",
     "svgo": "^2.8.0",
     "svgson": "^4.1.0"
   },
