+++ pnpm-workspace.yaml
@@ -1,3 +1,4 @@
 packages:
   - 'packages/*'
+  - 'tools/*'
   - 'site'
