+++ tools/outline-svg/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "outline-svg",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "start": "node ./main.mjs"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "minimist": "^1.2.6",
+    "svg-outline-stroke": "^1.3.1",
+    "svgson": "^5.2.1",
+    "tinypool": "^0.3.0"
+  }
+}
